bump: vllm version v0.13
- remove v0 codes and tests
- properly register attention backend
- applied some linting
