+++ tests/torch_compile/models/test_pooling_model_embed.py
@@ -81,7 +81,6 @@ def test_pooling_model_embed(
     env = {
         "VLLM_RBLN_USE_VLLM_MODEL": "1",
         "VLLM_DISABLE_COMPILE_CACHE": "1",
-        "VLLM_USE_V1": "1",
         "VLLM_RBLN_COMPILE_STRICT_MODE": "1",
     }
