+++ tests/torch_compile/models/test_pooling_model_score.py
@@ -91,7 +91,6 @@ def test_pooling_model_score(
     env = {
         "VLLM_RBLN_USE_VLLM_MODEL": "1",
         "VLLM_DISABLE_COMPILE_CACHE": "1",
-        "VLLM_USE_V1": "1",
         "VLLM_RBLN_COMPILE_STRICT_MODE": "1",
     }
