+++ vllm_rbln/v1/worker/utils.py
@@ -13,11 +13,13 @@
 # limitations under the License.
 """CPU affinity utilities for RBLN worker."""
 
+import math
 import os
 import platform
 from collections.abc import Callable
+from typing import Optional
 
-from vllm.config import ParallelConfig
+from vllm.config import ModelConfig, ParallelConfig
 from vllm.platforms import CpuArchEnum, current_platform
 from vllm.platforms.cpu import CpuPlatform, LogicalCPUInfo
 
@@ -25,6 +27,138 @@ import vllm_rbln.rbln_envs as envs
 from vllm_rbln.logger import init_logger
 
 logger = init_logger(__name__)
+
+
+# NOTE: This function comes from optimum-rbln. Keep in sync.
+def estimate_available_memory(
+    model_config: ModelConfig,
+    parallel_config: ParallelConfig,
+    nbits_per_param: Optional[int] = None,
+    n_model_params: Optional[int] = None,
+    kernel_size: Optional[int] = None,
+    buffer: Optional[int] = None,
+    num_runtimes: int = 2,
+    gpu_memory_utilization: float = 0.9,
+) -> int:
+    # We are finding max_num_blocks(x) that satisfies the following equation:
+
+    # available_dram - kernel_size - buffer
+    #     - num_layers * 2 * tensor_parallel_size
+    #     * align_2MB(
+    #         x
+    #         * block_size
+    #         * align_64(head_dim)
+    #         * math.ceil(num_key_value_heads / tensor_parallel_size)
+    #         * 2
+    #     ) > 0
+
+    # This inequality can be rewritten as follows:
+
+    # a - c * align_2MB(b * x) > 0
+    # where
+    #    a = available_dram - kernel_size - buffer
+    #    b = block_size
+    #         * align_64(head_dim)
+    #         * math.ceil(num_key_value_heads / tensor_parallel_size) * 2
+    #    c = num_layers * 2 * tensor_parallel_size
+
+    # We can rewrite the inequality as follows:
+    # k > align_2MB(b*x)
+    # where
+    #    k = a / c
+
+    # After that, we can derive the following equation:
+    # x = floor(2**21 / b * floor((k - 1) / 2**21))
+
+    def align(x: int, nbytes: int) -> int:
+        return int(math.ceil(x / nbytes) * nbytes)
+
+    def align_2MB(x: int) -> int:
+        return align(x, 2**21)
+
+    num_layers = model_config.get_num_layers(parallel_config)
+    vocab_size = model_config.get_vocab_size()
+    hidden_size = model_config.get_hidden_size()
+    tp_size = parallel_config.tensor_parallel_size
+    rsd_size = envs.VLLM_RBLN_TP_SIZE
+
+    # FIXME: Do we need to account for these configs?
+    # head_dim = model_config.get_head_size()
+    # num_key_value_heads = model_config.get_num_kv_heads(parallel_config)
+    # dp_size = parallel_config.data_parallel_size
+    # ep = parallel_config.enable_expert_parallel
+
+    # TODO(jongho): Update if target npu is REBEL.
+
+    device_name = current_platform.get_device_name().lower()
+    assert "rbln" in device_name
+    if "ca" in device_name:
+        # ATOM - RBLN-CA[xxx]
+        # ATOM DRAM - 16GB (single chip)
+        ATOM_DRAM_NBYTES = 16 * 2**30
+        ATOM_SYS_DRAM_NBYTES = 288 * 2**20
+        available_dram_bytes = rsd_size * (ATOM_DRAM_NBYTES -
+                                           ATOM_SYS_DRAM_NBYTES)
+        # ATOM - basic data type fp16
+        default_bits_per_param = 16
+    elif "cr" in device_name:
+        assert rsd_size == 1
+        # REBEL - RBLN-CR[xxx]
+        # REBEL DRAM - 144GB (quad chips, chiplet) - system(4G) = 140GB
+        REBEL_DRAM_NBYTES = 144 * 2**30
+        REBEL_SYS_DRAM_NBYTES = 4 * 2**30
+        REBEL_DRAM_NBYTES -= REBEL_SYS_DRAM_NBYTES
+        available_dram_bytes = REBEL_DRAM_NBYTES
+        # FIXME(RBLN) - basic data type fp8 for REBEL, for now fp16
+        default_bits_per_param = 16
+    else:
+        raise ValueError(
+            "invalid RBLN architecture, candidates = [ATOM(ca), REBEL(cr)]")
+
+    available_dram_bytes = int(available_dram_bytes * gpu_memory_utilization)
+
+    def check_oom(available_dram_bytes: int) -> None:
+        if available_dram_bytes <= 0:
+            raise MemoryError("Insufficient DRAM during block calculation. "
+                              "Try reducing gpu_memory_utilization.")
+
+    if kernel_size is None:
+        if n_model_params is None:
+            raise ValueError("`n_model_params` should be specified \
+                to estimate the kernel memory.")
+        # Get estimated kernel size (approximated)
+        # kernel_size
+        # - QKV params     - model parallel (tp) sharded
+        # - MLP or expert  - model parallel (ep) sharded
+        # - word embedding - non sharded,
+        #                    not included into device,
+        #                    hidden_size * vocab_size
+        # - lm head        - model parallel (tp) sharded,
+        #                    hidden_size * vocab_size
+        lm_heads_params = align(vocab_size, 64) * hidden_size
+        lm_heads_nbytes = (align_2MB(
+            lm_heads_params * default_bits_per_param // 8 / tp_size) * tp_size)
+        word_embedding_params = lm_heads_params
+        params = n_model_params - lm_heads_params - word_embedding_params
+        layer_nbytes = (align_2MB(params * nbits_per_param // 8 / num_layers) *
+                        num_layers)
+        kernel_size = layer_nbytes + lm_heads_nbytes
+    elif n_model_params is not None:
+        raise ValueError(
+            "Both `n_model_params` and `kernel_size` cannot be specified.")
+
+    available_dram_bytes -= kernel_size
+
+    if buffer is None:
+        # TODO: Accurate buffer estimation
+        buffer_per_runtime_per_core = 2**28  # 256MB per runtime
+        # 1 for prefill, 1 for decoder
+        buffer = buffer_per_runtime_per_core * num_runtimes
+    available_dram_bytes -= buffer
+
+    check_oom(available_dram_bytes)
+
+    return available_dram_bytes
 
 
 def get_autobind_cpu_ids(
